refactor(frontend): consolidate dashboard components
